net: Prevent node from binding to the same CService
Currently, if the user inadvertently starts the node with duplicate `bind` options,
such as `bind=0.0.0.0 -bind=0.0.0.0`, it will cause a fatal error with the
misleading message `Bitcoin Core is probably already running`.

This commit changes the behavior as follows:
If the duplicate binds are the same (e.g., `bind=0.0.0.0 -bind=0.0.0.0`), they will be ignored.
But if they are different (e.g., `bind=0.0.0.0:10080 -bind=0.0.0.0:10080=onion`), the
node will be terminated with a clear and specific message
(e.g., `Different bind settings assigned to address 0.0.0.0:10080`), since we cannot know
the user's true intent.